refactor: order-table 패키지 이동
