+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -2,8 +2,7 @@ package kitchenpos.order.application;
 
 import kitchenpos.application.FakeKitchenridersClient;
 import kitchenpos.order.domain.InMemoryOrderRepository;
-import kitchenpos.application.InMemoryOrderTableRepository;
-import kitchenpos.domain.*;
+import kitchenpos.ordertable.domain.InMemoryOrderTableRepository;
 import kitchenpos.menu.domain.InMemoryMenuRepository;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.order.domain.Order;
@@ -11,6 +10,8 @@ import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.OrderType;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
