+++ src/test/java/kitchenpos/order/domain/InMemoryOrderRepository.java
@@ -1,9 +1,6 @@
 package kitchenpos.order.domain;
 
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderRepository;
-import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTable;
 
 import java.util.*;
